menu view works (kinda)
