+++ menu_window.py
@@ -1,51 +1,51 @@
-import arcade
-import arcade.gui
-
-from screen_configuration import ScreenConfiguration
-
-# --- Method 1 for handling click events,
-# Create a child class.
-class QuitButton(arcade.gui.UIFlatButton):
-    def on_click(self, event: arcade.gui.UIOnClickEvent):
-        arcade.exit()
-
-
-class StartButton(arcade.gui.UIFlatButton):
-    def on_click(self, event: arcade.gui.UIOnClickEvent):
-        pass
-
-
-class MenuView(arcade.View):
-    def __init__(self, screen_config: ScreenConfiguration):
-        super().__init__()
-        self.configuration = screen_config
-        # --- Required for all code that uses UI element,
-        # a UIManager to handle the UI.
-        self.manager = arcade.gui.UIManager()
-        self.manager.enable()
-
-        # Set background color
-        arcade.set_background_color(arcade.color.DARK_BLUE_GRAY)
-
-        # Create a vertical BoxGroup to align buttons
-        self.v_box = arcade.gui.UIBoxLayout()
-
-        # Create the buttons
-        start_button = StartButton(text="Start Game", width=200)
-        self.v_box.add(start_button.with_space_around(bottom=20))
-
-        # Again, method 1. Use a child class to handle events.
-        quit_button = QuitButton(text="Quit", width=200)
-        self.v_box.add(quit_button)
-
-        # Create a widget to hold the v_box widget, that will center the buttons
-        self.manager.add(
-            arcade.gui.UIAnchorWidget(
-                anchor_x="center_x",
-                anchor_y="center_y",
-                child=self.v_box)
-        )
-
-    def on_draw(self):
-        self.clear()
-        self.manager.draw()
+# import arcade
+# import arcade.gui
+#
+# from screen_configuration import ScreenConfiguration
+#
+# # --- Method 1 for handling click events,
+# # Create a child class.
+# class QuitButton(arcade.gui.UIFlatButton):
+#     def on_click(self, event: arcade.gui.UIOnClickEvent):
+#         arcade.exit()
+#
+#
+# class StartButton(arcade.gui.UIFlatButton):
+#     def on_click(self, event: arcade.gui.UIOnClickEvent):
+#         pass
+#
+#
+# class MenuView(arcade.View):
+#     def __init__(self, screen_config: ScreenConfiguration):
+#         super().__init__()
+#         self.configuration = screen_config
+#         # --- Required for all code that uses UI element,
+#         # a UIManager to handle the UI.
+#         self.manager = arcade.gui.UIManager()
+#         self.manager.enable()
+#
+#         # Set background color
+#         arcade.set_background_color(arcade.color.DARK_BLUE_GRAY)
+#
+#         # Create a vertical BoxGroup to align buttons
+#         self.v_box = arcade.gui.UIBoxLayout()
+#
+#         # Create the buttons
+#         start_button = StartButton(text="Start Game", width=200)
+#         self.v_box.add(start_button.with_space_around(bottom=20))
+#
+#         # Again, method 1. Use a child class to handle events.
+#         quit_button = QuitButton(text="Quit", width=200)
+#         self.v_box.add(quit_button)
+#
+#         # Create a widget to hold the v_box widget, that will center the buttons
+#         self.manager.add(
+#             arcade.gui.UIAnchorWidget(
+#                 anchor_x="center_x",
+#                 anchor_y="center_y",
+#                 child=self.v_box)
+#         )
+#
+#     def on_draw(self):
+#         self.clear()
+#         self.manager.draw()
